requirements.txt
requests
python-dotenv
flask
flask-cors
pytest
pytest-mock
responses
